Moved static methods of MeteorEntity to new class MeteorUtils (such as the ones to spawn the meteors and stuff)
